+++ src/pushsource/_impl/backend/koji_source.py
@@ -14,6 +14,7 @@ from more_executors.futures import f_map
 from ..source import Source
 from ..model import RpmPushItem, ModuleMdPushItem
 from ..helpers import list_argument, try_int
+from .modulemd import Module
 
 LOG = logging.getLogger("pushsource")
 CACHE_LOCK = threading.RLock()
@@ -289,14 +290,41 @@ class KojiSource(Source):
             if self._module_filtered(file_path):
                 continue
 
-            # Possible TODO: koji also provides a checksum, which could be filled
-            # in here. However, it seems to be only MD5?
+            name = module["filename"]
+
+            # Set the push item name to NSVCA, if the modulemd file exists.
+            # NSVCA is a more meaningful name for the item than the modulemd.<arch>.txt,
+            # which will be the same for all modules of the same arch.
+            #
+            # Why only "if the modulemd file exists"? Well, it's only going to exist
+            # if the caller has mounted the NFS volume associated with koji. This will
+            # be true in production, and we could give a fatal error when this is not
+            # the case, but it seems like a very harsh requirement. e.g. most of the
+            # docs for this library are using fedora koji as an example, but nobody
+            # running outside of fedora infra will be able to mount that volume,
+            # meaning attempting to use the examples will immediately hit a brick wall.
+            #
+            # So, this is a compromise. If the file exists, it has to be valid, and we
+            # use it; this gives the best behavior in production. If the file doesn't
+            # exist, that's OK, and that's a compromise to avoid making the dev and test
+            # requirements too onerous.
+
+            if os.path.exists(file_path):
+                try:
+                    module = Module.from_file(file_path)
+                    name = module.nsvca
+                except:
+                    # If we fail, make it clear what we were attempting to do before raising
+                    LOG.exception(
+                        "In koji build %s, cannot load module metadata from %s",
+                        nvr,
+                        file_path,
+                    )
+                    raise
+
             out.append(
                 ModuleMdPushItem(
-                    name=module["filename"],
-                    src=file_path,
-                    dest=self._dest,
-                    build=meta["nvr"],
+                    name=name, src=file_path, dest=self._dest, build=meta["nvr"]
                 )
             )
         return out