+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `module_filter_filename` parameter to koji source, to select only a subset
   of modulemd files from a build (e.g. limit to certain arches).
 
+### Changed
+
+- `ModuleMdPushItem.name` now uses the NSVCA of a module rather than the filename of
+  a loaded modulemd file, when this metadata is available.
+
 ## [2.6.0] - 2021-05-27
 
 ### Added